MEjora en la lógica del caso de uso, testsdel caso de uso y ajustes en las dependencias del pom para que los tests funcionen

+++ src/main/java/com/zara/prices/application/domain/PricesRepository.java
@@ -4,5 +4,5 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public interface PricesRepository {
-    List<Prices> findPriceByProductDatesBrand(Integer productId, Long brandId, LocalDateTime date);
+    List<Prices> findPriceByProductDatesBrand(Long productId, Long brandId, LocalDateTime date);
 }
