+++ src/main/java/com/zara/prices/infraestructure/datasource/PricesJpaRepository.java
@@ -7,6 +7,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public interface PricesJpaRepository extends JpaRepository<PricesEntity, Long> {
-    List<PricesEntity> findByEndDateGreaterThanEqualAndStartDateLessThanEqualAndProductIdAndBrandId(LocalDateTime startDate, LocalDateTime endDate, Integer productId, Long brandId);
+    List<PricesEntity> findByEndDateGreaterThanEqualAndStartDateLessThanEqualAndProductIdAndBrandId(LocalDateTime startDate, LocalDateTime endDate, Long productId, Long brandId);
 
 }
